requirements.txt
pendulum==3.1.0
fastapi==0.104.1
uvicorn==0.24.0
prefect==2.14.5
neo4j==5.14.0
spacy==3.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0.tar.gz
requests==2.31.0
praw==7.7.1
newsapi-python==0.2.7
sec-api==1.0.16  # For SEC EDGAR
pypdf2==3.0.1  # For PDF extraction
structlog==23.2.0
python-dotenv==1.0.0
redis==5.0.1
pydantic==2.5.1
streamlit==1.28.2
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4
httpx==0.25.1  # For Grok API calls
tweepy==4.14.0  # For X API